fix: improve consistency with the app's theme
feat: change primary and secondary colors of the app to improve consistency with the app's theme
feat: change icon and label of the home and camera tabs to improve user experience
The changes made to the login, change password, and edit profile pages were to improve consistency with the app's theme. The primary and secondary colors of the app were also changed to improve consistency with the app's theme. The icon and label of the home and camera tabs were changed to improve user experience.

style(monitoring_page_widget.dart): update app bar styling and layout
refactor(monitoring_page_widget.dart): use FlutterFlowTheme for background color and text color
The app bar's preferred size has been reduced to 80.0 to improve the layout. The alignment of the back button has been changed to center it vertically. The background color of the scaffold and the text color have been updated to use the FlutterFlowTheme.

style: change color scheme of various widgets to match the new design
The changes involve updating the color scheme of various widgets to match the new design. Specifically, the background color of the RegisterWidget, SplashScreenWidget, and MyProfileWidget has been changed to a new tertiary color. The color of the app bar in these widgets has also been changed to a new shade of green. Additionally, the color of some icons and text has been changed to black to improve readability.